moved new helper to already exsited helper
