+++ frontend/src/app/core/helpers/questions-mandatory-helper.ts
@@ -16,3 +16,23 @@ export const getLocalMandatoryQuestions = (): IQuestion[] => [
         isMandatory: true,
     },
 ];
+
+export const addQuestionPlaceholder = (question: IQuestion): IQuestion => {
+    switch (question.order) {
+        case 0:
+            return {
+                ...question,
+                placeHolderText: 'Name Surname',
+            };
+        case 1:
+            return {
+                ...question,
+                placeHolderText: 'booker@email.com',
+            };
+        default:
+            return {
+                ...question,
+                placeHolderText: '',
+            };
+    }
+};
